Got the product form error message working
Created a view test, but then realized it's more of an integration test kinda deal.
Made changes to the form and controller
Updated the related controller test so it will pass, given the changes
Added a line to the integration test to test for the error message
THEN REALIZED I hadn't pulled the most recent changes and was working on something that had already been done. Well, crap. Here we go anyway.