SelectedOptions for LineItems. Styling.  Cart edit.
We now attach a selected option to a line item in order to track what the user wanted.  When re-ordering the same item with the same items, the cart will simply update quantity.  Same item with different options will create a new line_item.

Update button on cart now works.

There is a swimming fish! weeeee