Added carts/checkout.  Imrpoved checkout flow.
Now cart/edit => cart/checkout => cart/receipt.

Added views and checkout controller to cart.  Added sign in/out link to sidebar.  Added a bit to application controller to try to get devise to redirect back to it's referring page (as described here: https://github.com/plataformatec/devise/wiki/How-To:-redirect-to-a-specific-page-on-successful-sign-in ).  Don't think it works correctly.  Wonderful place for some capybara tests.